Support vue in sidebar
